Langchain/Chatbot using Paid and Open source LLM/requirements.txt
langchain
streamlit
langchain_openai
langchain_core
python-dotenv
langchain_community